Build fix
Canonical link: https://commits.webkit.org/43079@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@51652 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/svg/SVGListTraits.h
@@ -60,7 +60,7 @@ namespace WebCore {
 
     template<>
     struct SVGListTraits<false, float> {
-        static double nullItem() { return 0; }
+        static float nullItem() { return 0; }
         static bool isNull(double) { return false; }
     };
